% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.R
\name{acdc}
\alias{acdc}
\title{AdaCover Digraph Classifier}
\usage{
acdc(x, y, tau = 1, test_data, num_iter = 10, num_balls = 1)
}
\arguments{
\item{x}{Training data frame or matrix without labels}

\item{y}{Training labels}

\item{tau}{P-CCCD radii parameter between 0 and 1. Default is 1.}

\item{test_data}{Test data frame or matrix without labels}

\item{num_iter}{Number of iterations for the AdaBoost algorithm. Default is 10.}

\item{num_balls}{Cardinality of the covering ball set. Default is 1.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{final_predictions}{The final predictions on the test data}
     \item{alphas}{The alpha value at each stage of boosting}
}
}
\description{
Ensemble boosting method using FlexiBall classifiers as weak learners.
FlexiBalls are built using a similar framework as P-CCCD classifiers, but with
stochastic selection of covering balls and a parameter that controls the number
of balls used to cover the data. Boosted weights are used in the stochastic selection.
Multiclass classification uses SAMME framework.
}
\examples{

data("iris")

index <- sample(1:nrow(iris), nrow(iris) * 0.7)

x <- iris[index, 1:4]
y <- iris[index, 5]
test_data <- iris[-index, 1:4]

model <- acdc(x, y, tau = .5, test_data, num_iter = 5, num_balls = 1)

sum(model$final_predictions == iris[-index, 5])/nrow(iris[-index,])

}
\author{
Jordan Eckert
}
