% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCCCD.R
\name{aggregate_predictions}
\alias{aggregate_predictions}
\title{Aggregates final decision rule}
\usage{
aggregate_predictions(matrix_data)
}
\arguments{
\item{matrix_data}{matrix of predictions}
}
\value{
an list which includes:
 \item{predictions}{aggregated final predictions}
}
\description{
\code{aggregate_predictions} finds the most frequent class in each row
}
\details{
Maximum is determined by which.max in event of tie
}
\author{
Jordan Eckert
}
