% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESAMPLE.R
\name{classify_juggle}
\alias{classify_juggle}
\title{Resampled Pure and Proper Class Cover Catch Digraph Prediction}
\usage{
classify_juggle(object, newdata, type = "pred")
}
\arguments{
\item{object}{a \code{resampled_pcccd} object}

\item{newdata}{newdata as matrix or dataframe.}

\item{type}{"pred" or "prob". Default is "pred". "pred" is class estimations,
"prob" is \eqn{n\times k} matrix of class probabilities.}
}
\value{
a vector of class predictions (if type is "pred") or a \eqn{n\times p}
matrix of class probabilities (if type is "prob").
}
\description{
\code{classify_juggles} makes prediction using
\code{resampled_pcccd} object.
}
\examples{
n <- 1000
x1 <- runif(n, 1, 10)
x2 <- runif(n, 1, 10)
x <- cbind(x1, x2)
y <- as.factor(ifelse(3 < x1 & x1 < 7 & 3 < x2 & x2 < 7, "A", "B"))

# testing the performance
i_train <- sample(1:n, round(n*0.8))

x_train <- x[i_train,]
y_train <- y[i_train]

x_test <- x[-i_train,]
y_test <- y[-i_train]

juggle_pcccd <- juggle(x = x_train, y = y_train, n_model = 30)
pred2 <- classify_juggle(object = juggle_pcccd, newdata = x_test)

# confusion matrix
table(y_test, pred2)

# test accuracy
sum(y_test == pred2)/nrow(x_test)

}
\author{
Jordan Eckert jpe0018@auburn.edu
}
