% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METRICS.R
\name{cv52test}
\alias{cv52test}
\title{5 x 2 CV F-Test}
\usage{
cv52test(m1, m2)
}
\arguments{
\item{m1}{A matrix of the first model's predictions.}

\item{m2}{A matrix of the second model's predictions.}
}
\value{
a vector where the first value is the p-value for the F-test
and the next 10 values are the p-values for the t-tests.
}
\description{
\code{cv52test} calculates the 5 x 2 CV F-Test as well as the
10 t-tests for the difference between the two models.
}
\author{
Jordan Eckert
}
