% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCCCD.R
\name{dominate_greedy_matrix}
\alias{dominate_greedy_matrix}
\title{Greedy Domination Algorithm}
\usage{
dominate_greedy_matrix(A)
}
\arguments{
\item{A}{adjacency matrix of the cover.}
}
\value{
a list which includes:
 \item{i_dominant_list}{dominant sample indexes.}
}
\description{
\code{dominate_greedy_matrix} finds the MDS of a cover. Uses greedy
domination algorithm from West (2000).
}
\references{
Manukyan, A., & Ceyhan, E. (2016). Classification of imbalanced data with a
geometric digraph family. Journal of Machine Learning Research, 17(1),
6504–6543. https://jmlr.org/papers/volume17/15-604/15-604.pdf
}
\author{
Jordan Eckert
}
