% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.R
\name{dominate_stochastic_matrix}
\alias{dominate_stochastic_matrix}
\title{Stochastic selection for MDS}
\usage{
dominate_stochastic_matrix(M = M, num_balls, weights = weights_main)
}
\arguments{
\item{M}{Adjacency matrix of the digraph}

\item{num_balls}{Final cardinality of the MDS}

\item{weights}{Weights of the vertices}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{i_dominant_list}{index of the dominate set vertices}
}
}
\description{
Finds the minimum dominating set of a digraph greedily using AdaBoosted weights
}
\author{
Jordan Eckert
}
