% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCCCD.R
\name{find_hyperplane_boundary_points}
\alias{find_hyperplane_boundary_points}
\title{Creates region of proposed overlap}
\usage{
find_hyperplane_boundary_points(svm_model, x_pair, y_pair, delta_1, delta_2)
}
\arguments{
\item{svm_model}{SVM model with decision values}

\item{x_pair}{training data of the proposed overlap region}

\item{y_pair}{labels of the proposed overlap region}

\item{delta_1}{minimum distance for the first hyperplane}

\item{delta_2}{maximum distance for the second hyperplane}
}
\value{
a list which includes:
\item{indices}{index of points in overlap region}
\item{points}{coordinates of points in overlap region}
\item{labels}{labels of points in overlap region}
}
\description{
\code{find_hyperplane_boundary_points} uses SVM boundary and decision values to find the points between two parallel hyperplanes.
}
\details{
Fits a support vector machine using a Gaussian kernel to find the discrimination hyperplane.
Two parallel hyperplanes \eqn{H_1} and {H_2} are fit such that \eqn{d(H_1, H_2) = \delta_1 + \delta_2}.
Points inside region are considered in overlap region, if such region exists.
}
\author{
Jordan Eckert
}
