% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{generate_disjoint_square}
\alias{generate_disjoint_square}
\title{Generate datasets from disjointed Unit Square}
\usage{
generate_disjoint_square(n, m, d, Delta)
}
\arguments{
\item{n}{a numeric value. The size of class \eqn{X_0}.}

\item{m}{a numeric value. The size of class \eqn{X_1}.}

\item{d}{a numeric value. The dimension of the data.}

\item{Delta}{a numeric value. The overlap parameter between the two classes.}
}
\value{
a data frame with two classes.
}
\description{
\code{generate_disjoint_square} generates a data frame with two classes
from a unit square with varying overlap, but disjointed along the first feature,
that is \eqn{X_0 \sim U(0,1)^{d}} and \eqn{X_1 \sim U(1,2) \times U(\Delta, 1 + \Delta)^{d-1}}.
}
\examples{
library(plotly)
n <- 100
m <- 25
d <- 3
Delta <- 0

data <- generate_disjoint_square(n, m, d, Delta)
plot_ly(x = data$X1, y=data$X2, z=data$X3, type="scatter3d", mode="markers", color=data$Class)

}
\author{
Jordan Eckert
}
