% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{generate_normal_data}
\alias{generate_normal_data}
\title{Generate datasets from Normal distributions}
\usage{
generate_normal_data(n, m, d, dist)
}
\arguments{
\item{n}{a numeric value. The size of class \eqn{X_0}.}

\item{m}{a numeric value. The size of class \eqn{X_1}.}

\item{d}{a numeric value. The dimension of the data.}

\item{dist}{a numeric value. The Mahalanobis distance between the two classes.}
}
\value{
a data frame with two classes.
}
\description{
\code{generate_normal_data} generates a data frame with two classes
from Normal distributions for each feature with varying overlap,
that is \eqn{X_0 \sim N(0,1)^{d}} and \eqn{X_1 \sim N(\Delta, 1)^{d}} where
\eqn{\Delta} is a random direction some specified Mahalanobis distance away.
}
\examples{
library(plotly)
n <- 100
m <- 25
d <- 3
dist <- 1

data <- generate_normal_data(n, m, d, dist)
plot_ly(x = data$X1, y=data$X2, z=data$X3, type="scatter3d", mode="markers", color=data$Class)

}
\author{
Jordan Eckert
}
