% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{generate_unit_square}
\alias{generate_unit_square}
\title{Generate datasets from Unit Square}
\usage{
generate_unit_square(n, m, d, Delta)
}
\arguments{
\item{n}{a numeric value. The size of class \eqn{X_0}.}

\item{m}{a numeric value. The size of class \eqn{X_1}.}

\item{d}{a numeric value. The dimension of the data.}

\item{Delta}{a numeric value. The overlap parameter between the two classes.}
}
\value{
a data frame with two classes.
}
\description{
\code{generate_unit_square} generates a data frame with two classes
from a unit square with varying overlap, that is \eqn{X_0 \sim U(0,1)^d} and
\eqn{X_1 \sim U(\Delta, 1 + \Delta)^d}.
}
\examples{
n <- 50
m <- 25
d <- 2
Delta <- .5

x <- generate_unit_square(n, m, d, Delta)

plot(x[,c(1:2)], col = x$Class, asp = 1)

}
\author{
Jordan Eckert
}
