% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESAMPLE.R
\name{juggle}
\alias{juggle}
\title{Pure and Proper Class Cover Catch Digraph Resampled Classifier}
\usage{
juggle(
  x,
  y,
  n_model = 30,
  replace = TRUE,
  min_proportion = 0.5,
  tau = 1,
  verbose = FALSE
)
}
\arguments{
\item{x}{feature matrix or dataframe.}

\item{y}{class factor variable.}

\item{n_model}{an integer. Number of bagged iterations.}

\item{replace}{a bool. Should replacement be used in data sampling, if true, uses bootstrap sampling.}

\item{min_proportion}{a numeric. Minimum proportion of samples to use in each model, used only if replace is FALSE.}

\item{tau}{a numeric between \eqn{(0,1]}. Tau parameter for PCCCD.}

\item{verbose}{a bool. Should progress be printed to console.}
}
\value{
an object of "cccd_classifier" which includes:
 \item{i_dominant_list}{dominant sample indexes.}
 \item{x_dominant_list}{dominant samples from feature matrix, x}
 \item{radii_dominant_list}{Radiuses of the circle for dominant samples}
 \item{class_names}{class names}
 \item{k_class}{number of classes}
 \item{proportions}{proportions each class covered}
 \item{tau}{tau parameter for P-CCCD}
}
\description{
\code{juggle} fits a resampled Pure and Proper Class Cover Catch
Digraph (PCCCD) classification model.
}
\details{
Resampling framework for PCCCD.
}
\examples{
n <- 1000
x1 <- runif(n, 1, 10)
x2 <- runif(n, 1, 10)
x <- cbind(x1, x2)
y <- as.factor(ifelse(3 < x1 & x1 < 7 & 3 < x2 & x2 < 7, "A", "B"))

# testing the performance
i_train <- sample(1:n, round(n*0.8))

x_train <- x[i_train,]
y_train <- y[i_train]

x_test <- x[-i_train,]
y_test <- y[-i_train]

m_pcccd <- pcccd(x = x_train, y = y_train, tau = 1)
pred <- classify_pcccd(pcccd = m_pcccd, newdata = x_test)

juggle_pcccd <- juggle(x = x_train, y = y_train, n_model = 30)
pred2 <- classify_juggle(object = juggle_pcccd, newdata = x_test)

# confusion matrix
table(y_test, pred)
table(y_test, pred2)

# test accuracy
sum(y_test == pred)/nrow(x_test)
sum(y_test == pred2)/nrow(x_test)

}
\author{
Jordan Eckert
}
