% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.R
\name{weighted_pcccd}
\alias{weighted_pcccd}
\title{Weighted Pure and Proper Class Cover Catch Digraph Classifier}
\usage{
weighted_pcccd(x = x, y = y, weights = weights, num_balls, tau)
}
\arguments{
\item{x}{feature matrix or dataframe.}

\item{y}{class factor variable.}

\item{weights}{Weighting vector. Default is equal weighting.
Smaller numbers results in less likely to be used in dominating algorithm.}

\item{num_balls}{Number of minimum dominating balls to be used in the cover}

\item{tau}{PCCCD radii parameter between 0 and 1. Default is 1.}
}
\value{
an object of "cccd_classifier" which includes:
 \item{i_dominant_list}{dominant sample indexes.}
 \item{x_dominant_list}{dominant samples from feature matrix, x}
 \item{radii_dominant_list}{Radiuses of the circle for dominant samples}
 \item{weights}{boosted weights used to be returned}
 \item{class_names}{class names}
 \item{k_class}{number of classes}
 \item{proportions}{defunct parameter}
}
\description{
\code{weighted_pcccd} fits a Pure and Proper Class Cover Catch
Digraph (PCCCD) classification model using weights of points to greedily determine MDS
}
\author{
Jordan Eckert jordan@eckert.network
}
