% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCCCD.R
\name{svcccd}
\alias{svcccd}
\title{Support Vector Hybrid Class Cover Catch Digraph Classifier}
\usage{
svcccd(x, y, test_data, gamma = 0.01, e = 1, cost = TRUE)
}
\arguments{
\item{x}{feature matrix or dataframe.}

\item{y}{class factor variable.}

\item{test_data}{Test data frame or matrix without labels}

\item{gamma}{Parameter to control Gaussian SVM kernel}

\item{e}{Parameter to RW-CCCD}

\item{cost}{boolean. If TRUE, uses cost weighted SVM to determine boundary.}
}
\value{
an object of "svcccd" which includes:
 \item{final_predictions}{final predictions}
}
\description{
\code{svcccd} fits a support vector hybrid class cover catch digraph
classifier.
}
\details{
Fits a support vector machine using a Gaussian kernel to find a discrimination hyperplane.
Two parallel hyperplanes \eqn{H_1} and \eqn{H_2} are fit such that \eqn{d(H_1, H_2) = \delta}.
Points inside region are classified using RW-CCCD, if such region exists. If not, classification
proceeds using SVM.
}
\examples{
library(e1071)
data("iris")
set.seed(1234)
index <- sample(1:nrow(iris), nrow(iris) * 0.7)

x <- iris[index, 1:4]
y <- iris[index, 5]

test_data <- iris[-index, 1:4]
test_label <- iris[-index, 5]

gamma <- .01
e <- 1

model <- svcccd(x, y, test_data, gamma, e)
sum(model$final_predictions == test_label) / length(test_label)
}
\author{
Jordan Eckert
}
